% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalExprOnPartitionings.R
\name{evalExprOnPartitionings}
\alias{evalExprOnPartitionings}
\title{Applies dynamic evaluation on two partitionings}
\usage{
evalExprOnPartitionings(p1 = NULL, n1 = -1, p2 = NULL, n2 = -1, expr)
}
\arguments{
\item{p1}{the first list of observations}

\item{n1}{the number of partitions for p1}

\item{p2}{the second list of observations}

\item{n2}{the number of partitions for p2}

\item{expr}{the expression to be evaluated}
}
\description{
allows to pass an expression that will be evaluated, employing two partitioning arguments that will be converted from lists of observations
}
