\name{ocDte-package}
\alias{ocDte-package}
\alias{ocDte}
\docType{package}
\title{
\packageTitle{ocDte}
}
\description{
\packageDescription{ocDte}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ocDte}
\packageIndices{ocDte}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ocDte}

Maintainer: \packageMaintainer{ocDte}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
