% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning.R
\name{partitioning}
\alias{partitioning}
\title{A converter function for partitionings}
\usage{
partitioning(list = NULL, num = -1)
}
\arguments{
\item{list}{the list of values to be partitioned}

\item{num}{the number of partitions to create}
}
\description{
creates a partitioning from a list of values
}
