% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOperationOnPartitionings.R
\name{runOperationOnPartitionings}
\alias{runOperationOnPartitionings}
\title{Runs operations on partitionings}
\usage{
runOperationOnPartitionings(
  op = Null,
  p1 = NULL,
  n1 = -1,
  p2 = NULL,
  n2 = -1,
  options = NULL
)
}
\arguments{
\item{op}{the name of the operation}

\item{p1}{the first list of observations}

\item{n1}{the number of partitions for p1}

\item{p2}{the second list of observations}

\item{n2}{the number of partitions for p2}

\item{expr}{the expression to be evaluated}
}
\description{
allows to specify an operation that will be evaluated, employing two partitioning arguments that will be converted from lists of observations. In openCPU there are currently issues with the lexical scope of eval(), therefore we employ this less dynamic workaround...
}
